refactor: split useFilteredPlayers and normalize player data
